refact: rename paging widget package
