+++ instantsearch-compose/src/main/java/com/algolia/instantsearch/compose/filter/facet/internal/FacetListConnectionPager.kt
@@ -1,6 +1,6 @@
 package com.algolia.instantsearch.compose.filter.facet.internal
 
-import com.algolia.instantsearch.compose.paging.Paginator
+import com.algolia.instantsearch.compose.list.Paginator
 import com.algolia.instantsearch.core.connection.ConnectionImpl
 import com.algolia.instantsearch.helper.filter.facet.FacetListViewModel
 import com.algolia.search.model.search.Facet
