+++ instantsearch-compose/src/main/java/com/algolia/instantsearch/compose/searchbox/SearchBoxConnector.kt
@@ -1,6 +1,6 @@
 package com.algolia.instantsearch.compose.searchbox
 
-import com.algolia.instantsearch.compose.paging.Paginator
+import com.algolia.instantsearch.compose.list.Paginator
 import com.algolia.instantsearch.compose.searchbox.internal.SearchBoxConnectionPaginator
 import com.algolia.instantsearch.core.connection.Connection
 import com.algolia.instantsearch.core.searchbox.SearchBoxViewModel
