+++ instantsearch-compose/src/main/java/com/algolia/instantsearch/compose/searchbox/internal/SearchBoxConnectionPaginator.kt
@@ -1,6 +1,6 @@
 package com.algolia.instantsearch.compose.searchbox.internal
 
-import com.algolia.instantsearch.compose.paging.Paginator
+import com.algolia.instantsearch.compose.list.Paginator
 import com.algolia.instantsearch.core.connection.ConnectionImpl
 import com.algolia.instantsearch.core.searchbox.SearchBoxViewModel
 import com.algolia.instantsearch.helper.searchbox.SearchMode
